Use state spread for NamespaceKPIs setState
If we ever add fields to the state object, we'll need to include them
in every call to setState, since it overwrites the entire state.
So instead of having to manually add them to every setState call,
we can just spread the current state so you only need to specify
the changing fields afterwards. This makes it more maintainable
if you need to add a state field and makes this pattern ubiquitous
throughout the file.