Refactor EditFacts for form validtion
This refactor separately tracks fact values and form field values (which
are now different), and creates transformations to normalize both
values to check if they are equal. This is required because the new
datetime-local fields and numeric input updates do not exactly match
the value of the fact in the db any longer.

This also ensures the form field is disabled if the fact type is
read-only