Add Markdown-enabled form field
This is a text area with a toggle that will render the markdown content.
I had to add `className` support to the TextArea component for the UI to
work as one would expect it to -- it was the only good way to force the
two components to have the same size.