Don't always update ops log recorded_at on edit
This fixes a bug where the recored_at value (and also created_at,
if it was defined) was always replaced in the patch request. This
was because the format of the db value and the format from
toISOString() are slightly different (+00:00 vs .000Z), and the
json patch compare would always consider them different, even
if they were the same actual datetime. This would make an erroneous
request and not actually change any data. To prevent this, this
commit normalized the db value and field value to be the same.