Add OpsLog activity feed entries
This adds a new type of entry to the activity feed on the dashboard
page for Ops Logs. Ops Logs have nullable fields, so the displayed
text varies depending on which fields have values. Ops Log entry
list items are also links to the ops log page where you can view
that entry.