Pass response headers to httpGet success callback
This refactor modifies the httpGet method's onSuccess callback
signature, so instead of taking only the response payload ("data")
as argument, it takes an object that includes both "data" and now
also "headers". This allows the user to have access to the response
headers, which will be required for following Link header URLs. It
also has the added benefits of enforcing the argument names on the
users - some used "result", etc., and now must all be "data" - and
is extensible if we need to add more fields.