Listen for key down events on SlideOver
This is to enable using arrow keys to navigate through rows
on a table. This gives SlideOver an optional onKeyDown prop that
the caller can pass in if they need to hook into key down events.
Some manipulations of the SlideOver content can make the relevant
div lose focus and no longer track events. The caller can manually
re-enable focus by passing in a different focusTrigger prop when
this occurs.