Use arrow keys to nav thru ops log table
This adds the ability to use the left and right arrow keys to
move through entries in the ops log table. It updates the highlighted
row and the v query param as you navigate. It does not navigate
if you are either editing the current entry or have the delete
confirmation dialog open.