Add arrow icons when ops log is navigable
This adds arrow icons to the top of the SlideOver when viewing
an ops log if the user has the ability to navigate thru the table
via the arrow keys. This is because they can't (as of now) if you
navigate to /ui/operations-log?v=42 by itself - it only works if
you click on a row since that is what sets the selected index.