Allow arrow nav when linked to ops log page
This allows users to see the arrow icons and use their arrow
keys to navigate around the ops log table entries, even if they
are navigating to the URL directly. This is because we check for
this case and set the selectedIndex if the ops log entry with the
v parameter is in the table. In the unlikely, but possible, scenario
where they have filtered the table to exclude this value but
manually added the v param (or vice versa), they won't get arrows,
but will still view the correct ops log and have the correctly-
filtered table