Fix editing ops log entries with projects
This fixes an error where you could not edit the ops log if it had
a project set. This bug was introduced in fc6fa1c when
removing the chaining operator in front of .trim(), as the value
project_id was actually an integer and had no trim(). In fixing it,
I refactored the names of some things to make it more clear that
the fieldValues state holds values for the input fields (in their
natural types, e.g. string for option), and the values() function
was renamed to toSchemaValues(fieldValues) to indicate that it
transforms field values to ops log schema values (in their natural
types, e.g. project_id is an int) for validation and PATCHing.