Let users click ops log navigation arrows
This lets users click the left and right arrows when viewing an
ops log so they can navigate through the table just like using
the arrow keys. It displays a pointer cursor if they can click it
and it disables highlighting because if you click it it will end up
highlighting the entire title, which is not desired.