Make ops log left/right arrows keyboard-accessible
This wraps the left and right arrow icons in buttons so we get
highlighting, tab navigation, and trigger on "Enter" keypress for
free. We also don't allow tab navigation if the arrow is grayed-
out, since the user can't press it. There are a few style changes
in here, which just maintain the current vertically-aligned style,
and enforce the same size and offset blue ring on focus.