Handle syntax errors on project search submit
Fixes a bug where having a syntax error in the filter on the projects
search page results in the page locking up due to an unhandled
exception. This is easily reproduced by omitting an "AND" in a 2-part
filter, or similar. This commit wraps the filter parsing in a try/catch,
surfacing the error to the user via the error message alert bar.

NOTE: Since we get the error messaging directly from the Error instance
thrown from @cybernetex/kbn-es-query, we have limited ability to
localize the wording. I've opted for simplicity, taking the message as
is, but that means it is only in English. Alternately, we could instead
show a generic but localized message. Offering more helpful, localized
messaging would require either some more parsing of the error message,
or a parsing library that surfaces more error details than this one
seems to.

Fixes #47