Rewrite Project.Create to use new automations
I left the remnants of the previous implementation in place to minimize
the churn in this commit. The automations code has been moved completely
to the server side now. The /ui/available-automations API returns the
list of automations available for THIS user and the selected project
type. We simply display toggles and pass the selected automations back
to Imbi.

I also removed the ability to create links, select environments, and set
URLs from this screen. Most of the links are going to be set via
automations in the future since having incorrect links is a problem in
current usage. You can easily add them later and I expect that they are
more likely to be correct ;)