Simplify state when fetching SSM params
onFetch was being used as a proxy for ensuring the effect loaded
only once. Instead, it should just use the empty array as its
dependency list as this is the React way to do things. There is
also no need to check fetching at the top either, since this can
only be run once.