Add Component Usage report
I implemented the vast majority of this in a new component named Report
based on the Namespace KPIs report. The textual content for the report
help section comes from a prefixed section of the localization catalog.
In this case `reports.componentUsage` which contains a `title` and a
property under `columns` for each of the columns. The result is that
all of the text is in the localization catalog and the ComponentUsage
report view is less than 30 lines.

The next commit will migrate the NamespaceKPIs view to use the new
component.