Simplify a bunch of wonky height management
We were doing some now-unnecessary height manipulation (now that
we have a grid). This ends up simplifying things so we don't need
to pass around whether we're editing or not and adjusting the height
accordingly. That had issues anyway where different cards would
shrink when you edited something else. I also simplified the max
height of the project feed, instead of doing any calculations just
cap it at the screen height, which is reasonable enough and less
code to understand.