Prepare for yarn-modern
Setting the packageManager prop in package.json to the old version of
yarn ensures that we continue to get what we want. Yarn Classic (1.x)
will ignore this. However, if you enable `corepack` and use a modern
version of yarn, it will install and use the specified version. So,
I'm going to set it to the newest in the Yarn Classic line for the
time being. If we decide to adopt Yarn Modern with all of it's goodness,
then we can change the version number. Did I mention that Yarn Modern
doesn't require node_modules ;)

https://yarnpkg.com/migration/guide