Fix multi-page fetching in ComponentList
I added the `fetchPages` function in utils to make this easy to do just
about anywhere. The implementation in the view was broken due to JS
variable scoping *and* the fact that `httpGet` is sync but async :/

The `fetchPages` function is built on top of `httpGet` so it has the
same weirdness wrto. sync/async. It is best to think of it as a
background continuation since it uses `setTimeout` to queue up the next
page fetch. It invokes a callback when processing API responses and
includes a "done" flag so you can keep track of whether it is processing
or not at the component level.

See the `ComponentList` for the expected usage pattern.