Add sorting to NavigableTable
This is a single column sort which is a lot easier to implement and
understand (from the user perspective) than the multicolumn sort
implemented in a few other places. I also included the option to specify
the default sort as a property so that the UI can be sorted and make
sense on the initial render.