Add UI for updating existing params
This adds an editing view for params at a specific name. It allows
users to update the name of the param, in addition to the types
and values in each environment.

Behind the scenes, there are two flows for changing the name vs
only changing types & values. If you change the name, since this
isn't a patch on the same endpoint, we make two requests, one to
delete the existing params and another to post the new ones. This
mirrors what we need to do with AWS SSM anyway, since you can't
"update" the name. If you don't change the name and only change
the type and/or values, we perform a patch on that name. Users
can independently update type, value, add new params in environments
if they don't already exist, or even delete params from here if
they want.

Frontend validation here ensures that users need to include both a
type and a value for any param added, otherwise they can't click
"Save".

Error messages in the response are shown to the user in the slide
over.