Cap ProjectFeed height at tallest sibling
The new outer div acts as a layout container that is exactly the
height of the tallest sibling. The height: 0px and min-height: 100%
trick enforces this because the height will make it ignore its
own content, and then the min-height will take on the height of
the parent, which is determined by the grid. Since the min-height
is larger than the height, it will always be selected as the actual
height of the element.

This fixes the height problems, so that a small feed will only
take up as much space as needed, but a large feed will be capped
at the largest sibling height.

However, this breaks scrolling the feed because the height of the
parent is no longer explicit. To fix this, we switch the parent
to a column flexbox to get scrolling for free.