Added missing port to webpack config
I think that this changed when I switched to react-refresh in 827cc453.
Everything seemed to work but there was a warning/error in the console
about the websocket being disconnected on `ws://localhost:8000/ws`. I
noticed that my imbi api was receiving requests for `/ws`. Explicitly
setting the port to 8080 fixed it.