Fix the global search box on project search page
Entering a search into the *global* search box navigates to the project
search page with the filter added to the URL. Doing this while *on the
project search page* would update the URL but not the filter since the
query params are only read on initial render (eg, `useCallback({}, [])`).

Adding a small snippet that sets the filter & refresh state when the
location is updated and the f search param is included and does not
match the current filter seems to address the issue.