Make ProjectPicker value an object ALWAYS!
The final state of the picker used to be the numeric value of the
project ID or null. This didn't work so well when providing the value
from the URL in NewEntry so I converted it to always be an object.

The ProjectPicker initialization ensures that the namespace, project
type, and project IDs are congruent. It also takes care of converting
between numbers and strings were necessary ... most notably when
assigning a value to the <select> boxes.